+++ task_manager/templates/tasks/list.html
@@ -36,7 +36,7 @@
         {% else %}
             <td></td>
         {% endif %}
-        <td>{{ task.created_at }}</td>
+        <td>{{ task.created }}</td>
         <td>
           <a href="{% url 'task_update' task.id %}" class="btn btn-primary btn-sm">{% translate 'Update' %}</a>
             <a href="{% url 'task_delete' task.id %}" class="btn btn-danger btn-sm">{% translate 'Delete' %}</a>
